feature: 요구사항 6 - component scan refactor: util class 확장
